Move locale to the state + reducer
