+++ src/i18n/index.js
@@ -1,3 +1,6 @@
-export * from './TranslationLoader';
+export const DEFAULT_LOCALE = 'en';
+
+export * from './TranslationUtils';
 export Translate from './Translate';
 export TranslationProvider from './TranslationProvider';
+export TranslationReducer from './TranslationReducer';
